Create DatasetDelimeter.sas
Delimiters and how to change them: (what if your data is separated by a comma, dot, etc.?)

Delimiters are used to separate values. this can a b;lank space, a comma, or even a dot.

In the code above we have assumed that the current delimiter is a blank space so the use of the delimiter function was unnecessary (look to previous file on data set options). but someone, dunno who, has changed the data in the text file so that now, the data is separated by a dot -> "1997.19000".

The use of the DLM function tells the compiler that there has been a changed and to watch out for the delimiter "."